fix: moved UsageCounter to separate component modlet out of ChatContent scope, removed comments
